requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
python-pptx>=0.6.21
python-docx>=1.0.0
